ACM-21895: Address smcavey review — policy builder additionalColumns + test reorganization
1. Add additionalColumns ...ExtractProperty to ConfigPolicyResourceBuilder,
   CertPolicyResourceBuilder, and OperatorPolicyResourceBuilder (CRDs that
   may have additionalPrinterColumns defined). Update transformer.go to pass
   event.AdditionalPrinterColumns... at all three call sites.

2. Move test helpers (setupTransformConfig, setupConditionsConfig) and all
   specific-kind CollectorConfig integration tests from the ad-hoc
   refactored_transforms_test.go into genericResource_test.go, which is
   the established home for generic/configurable-collection test coverage.
   Delete refactored_transforms_test.go.
