+++ pkg/transforms/policy.go
@@ -92,7 +92,7 @@ func getPolicyCommonProperties(c *unstructured.Unstructured, node Node) Node {
 	return node
 }
 
-func OperatorPolicyResourceBuilder(c *unstructured.Unstructured) *PolicyResource {
+func OperatorPolicyResourceBuilder(c *unstructured.Unstructured, additionalColumns ...ExtractProperty) *PolicyResource {
 	node := transformCommon(c) // Start off with the common properties
 	node = getPolicyCommonProperties(c, node)
 	node = recordRelatedObjects(c, node)
@@ -136,18 +136,18 @@ func OperatorPolicyResourceBuilder(c *unstructured.Unstructured) *PolicyResource
 	node.Properties["deploymentAvailable"] = strconv.FormatBool(deploymentAvailable)
 	node.Properties["upgradeAvailable"] = strconv.FormatBool(upgradeAvailable)
 
-	node = applyDefaultTransformConfig(node, c)
+	node = applyDefaultTransformConfig(node, c, additionalColumns...)
 	return &PolicyResource{
 		node: node,
 	}
 }
 
-func ConfigPolicyResourceBuilder(c *unstructured.Unstructured) *PolicyResource {
+func ConfigPolicyResourceBuilder(c *unstructured.Unstructured, additionalColumns ...ExtractProperty) *PolicyResource {
 	node := transformCommon(c) // Start off with the common properties
 	node = getPolicyCommonProperties(c, node)
 	node = recordRelatedObjects(c, node)
 
-	node = applyDefaultTransformConfig(node, c)
+	node = applyDefaultTransformConfig(node, c, additionalColumns...)
 	return &PolicyResource{
 		node: node,
 	}
@@ -233,13 +233,13 @@ func parseConfigPolicyRelatedObject(item any) *relatedObject {
 	return obj
 }
 
-func CertPolicyResourceBuilder(c *unstructured.Unstructured) *PolicyResource {
+func CertPolicyResourceBuilder(c *unstructured.Unstructured, additionalColumns ...ExtractProperty) *PolicyResource {
 	node := transformCommon(c) // Start off with the common properties
 
 	detailMap, found, err := unstructured.NestedMap(c.Object, "status", "compliancyDetails")
 	if len(detailMap) == 0 || !found || err != nil {
 		node = getPolicyCommonProperties(c, node)
-		node = applyDefaultTransformConfig(node, c)
+		node = applyDefaultTransformConfig(node, c, additionalColumns...)
 		return &PolicyResource{node: node}
 	}
 
@@ -286,7 +286,7 @@ func CertPolicyResourceBuilder(c *unstructured.Unstructured) *PolicyResource {
 	node.Metadata["relObjs"] = certList
 
 	node = getPolicyCommonProperties(c, node)
-	node = applyDefaultTransformConfig(node, c)
+	node = applyDefaultTransformConfig(node, c, additionalColumns...)
 	return &PolicyResource{
 		node: node,
 	}
